requirements.txt
discord.py
python-dotenv
neuralintents
